+++ test/brod_client_SUITE.erl
@@ -43,6 +43,7 @@
         , t_sasl_plain_file_ssl/1
         , t_sasl_callback/1
         , t_magic_version/1
+        , t_get_partitions_count_safe/1
         ]).
 
 -include_lib("common_test/include/ct.hrl").
@@ -111,6 +112,17 @@ all() ->
 
 
 %%%_* Test functions ===========================================================
+
+t_get_partitions_count_safe(Config) when is_list(Config) ->
+  Client = ?FUNCTION_NAME,
+  ok = start_client(?HOSTS, Client),
+  Topic = <<"unknown-topic-001">>,
+  Res = brod:get_partitions_count_safe(Client, Topic),
+  ?assertMatch({error, unknown_topic_or_partition}, Res),
+  %% expect the 'unknown' result to be cached
+  Res2 = brod_client:lookup_partitions_count_cache(Client, Topic),
+  ?assertMatch({error, unknown_topic_or_partition}, Res2),
+  ok = brod:stop_client(Client).
 
 t_skip_unreachable_endpoint(Config) when is_list(Config) ->
   Client = t_skip_unreachable_endpoint,